Ajout d'un nouveau fichier dans l'arborescence projet : uTest.sh (à la racine du projet). Ce fichier sert de point de point d'entrée au buildbot pour executer tous les tests unitaires.
Le résultat de ces tests est publié sous test/output sous forme de fichiers xml très détaillés.

git-svn-id: svn://team-ard.com/ARP/trunk@864 c9b90f90-cbc1-11df-9d65-af01d29d43da

+++ arp_hml/src/orocos/can/CanOpenNode.hpp
@@ -58,8 +58,8 @@ namespace arp_hml
         /** CAN adress of the node */
         nodeID_t propNodeId;
 
-        /** Timeout before considering a node is not responding to a NMT request (in ms) */
-        int propNmtTimeout;
+        /** Timeout before considering a node is not responding to a NMT request (in s) */
+        double propNmtTimeout;
 
         /** name of the CanOpenController this component will connect*/
         string propCanOpenControllerName;
